Added parameter "mode_command_template"
This parameter allow to implement template to transform payload messages sent via MQTT to climate device, in order to let manage also those devices that does not support native modes of HASS (off,heat,cool), so a "cool" message via this parameter can be transformed to whatever message you need and sent out. For example on my thermostat there is an "Holiday" mode that normally is not manageable by HA apart by building a custon component. With this addition, i can use, for example the "dry" mode of climat control, transform it to the message "holiday" before sending it out to my thermostat, and have on Lovelace page display the "dry" mode with a custon name and icon as "Holiday Mode".
Same implementation can be done for templating every outgoing message, not only for received ones.